Created update-apt-vers.sh
This shell executable script will check for dependency issues, check for updates as well as update packages, remove packages that are no longer needed and also remove configuration files that are used only by the no longer needed package, as well as do an automatic clean up of the temp files that were downloaded for updating the packages.  This shell script requires that "apt" has been already installed on the Debian system.  

This script does require user input of "Y" or "N" as prompts are not ignored during the package updating process.